# specs/003-catalog-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Catalog API
  version: 1.2.0
  description: >
    Public, read-only catalog API (Feature 003). Single-tenant: every clinic value comes from the
    database. Field names and types match frontend/src/types/content.ts; fields marked
    "additive" are new and optional for existing clients. All prices are whole PKR. Schedule
    times are wall-clock "HH:MM" in the clinic's timeZone.

    Every response carries X-Request-ID and security headers. Catalog 200 responses carry
    ETag and Cache-Control: public, max-age=300; a matching If-None-Match returns 304.
    Every /api/v1 endpoint is rate limited per IP (default 60/min) and may return 429.

    Availability and booking (Feature 005). Instants are UTC ISO 8601 ("Z"); every instant is
    accompanied by localDate / localTime in the clinic timeZone. Slot and booking responses are
    never cached (Cache-Control: no-store). POST /appointments is a server-to-server call: it
    requires X-Proxy-Secret, and X-Client-IP is honoured only with a valid secret.

    Command Centre (Feature 006). Every /api/v1/admin/* operation (tag command-centre) is called only by the
    website server: it requires X-Proxy-Secret, all but sign-in and demo start also require X-Session-Token,
    and state-changing operations (and POST /admin/bookings/search) require X-CSRF-Token. Responses are
    Cache-Control: no-store and X-Robots-Tag: noindex. The access policy of each operation is `x-policy`
...
servers:
...
paths:
...
  /api/v1/admin/auth/sign-in:
    post:
      tags: [command-centre]
      operationId: adminSignIn
      x-policy: PUBLIC_PROXY
      security: [ { proxySecret: [] } ]
      description: Ends any presented session first (fixation). Limits per IP (429 rate_limited) and per typed email (429 account_locked).
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SignInRequest" } } } }
      responses:
        "200": { description: signed in, content: { application/json: { schema: { $ref: "#/components/schemas/SessionIssued" } } } }
        "401": { description: "sign_in_failed (generic)", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
        "429": { $ref: "#/components/responses/AdminRateLimited" }
  /api/v1/admin/demo/start:
    post:
      tags: [command-centre]
      operationId: adminDemoStart
      x-policy: PUBLIC_PROXY
      security: [ { proxySecret: [] } ]
      description: Creates a 2-hour read-only demo session for today's Karachi date. Ends any presented session first.
      responses:
        "200": { description: demo session, content: { application/json: { schema: { $ref: "#/components/schemas/SessionIssued" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/AdminRateLimited" }
  /api/v1/admin/auth/me:
    get:
      tags: [command-centre]
      operationId: adminMe
      x-policy: SELF
      security: [ { proxySecret: [], sessionToken: [] } ]
      responses:
        "200": { description: current viewer, content: { application/json: { schema: { $ref: "#/components/schemas/Viewer" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /api/v1/admin/auth/sign-out:
    post:
      tags: [command-centre]
      operationId: adminSignOut
      x-policy: SELF
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      responses:
        "204": { description: session ended }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/auth/change-password:
    post:
      tags: [command-centre]
      operationId: adminChangePassword
      x-policy: SELF_STAFF
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      description: Ends all of the staff member's sessions and issues a fresh one.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ChangePasswordRequest" } } } }
      responses:
        "200": { description: new session, content: { application/json: { schema: { $ref: "#/components/schemas/SessionIssued" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
...
  /api/v1/admin/bookings/search:
    post:
      tags: [command-centre]
      operationId: adminBookingSearch
      x-policy: READ
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      description: A read sent as POST so personal search terms never appear in a URL (FR-021). Default = today, all doctors, all statuses.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/BookingSearchRequest" } } } }
      responses:
        "200": { description: page of bookings, content: { application/json: { schema: { $ref: "#/components/schemas/BookingPage" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/admin/bookings/{reference}:
    get:
      tags: [command-centre]
      operationId: adminBookingDetail
      x-policy: READ
      security: [ { proxySecret: [], sessionToken: [] } ]
      parameters: [ { $ref: "#/components/parameters/Reference" } ]
      responses:
        "200": { description: detail (phone and email masked), content: { application/json: { schema: { $ref: "#/components/schemas/BookingDetail" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: not_found }
  /api/v1/admin/bookings/{reference}/status:
    post:
      tags: [command-centre]
      operationId: adminBookingChangeStatus
      x-policy: WRITE
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      parameters: [ { $ref: "#/components/parameters/Reference" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/StatusChangeRequest" } } } }
      responses:
        "200": { description: changed (booking + status history + audit in one transaction), content: { application/json: { schema: { $ref: "#/components/schemas/StatusChangeResult" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: not_found }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/admin/bookings/{reference}/status/undo:
    post:
      tags: [command-centre]
      operationId: adminBookingUndoStatus
      x-policy: WRITE
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      parameters: [ { $ref: "#/components/parameters/Reference" } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UndoRequest" } } } }
      responses:
        "200": { description: restored, content: { application/json: { schema: { $ref: "#/components/schemas/BookingDetail" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: "undo_unavailable / slot_taken", content: { application/json: { schema: { $ref: "#/components/schemas/BookingChanged" } } } }
  /api/v1/admin/bookings/{reference}/reveal-phone:
    post:
      tags: [command-centre]
      operationId: adminBookingRevealPhone
      x-policy: READ
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      description: Staff → audit row booking.phone_revealed (reference only). Demo → synthetic number, nothing stored.
      parameters: [ { $ref: "#/components/parameters/Reference" } ]
      responses:
        "200": { description: full phone, content: { application/json: { schema: { $ref: "#/components/schemas/PhoneReveal" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: not_found }
...
components:
  securitySchemes:
    proxySecret: { type: apiKey, in: header, name: X-Proxy-Secret }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
...
    Reference:
      name: reference
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{10}$" }
      description: Booking reference without dash (Crockford base32).
...
  responses:
...
    Unauthenticated: { description: "401 not_signed_in / session_expired", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Forbidden: { description: "403 forbidden / demo_read_only / password_change_required / csrf_failed", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Conflict: { description: "409 (see code)", content: { application/json: { schema: { $ref: "#/components/schemas/BookingChanged" } } } }
    Validation: { description: "422 validation_error / weak_password", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    AdminRateLimited: { description: "429 rate_limited / account_locked (Retry-After header)", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  schemas:
...
    ErrorInfo:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, enum: [not_found, method_not_allowed, validation_error, rate_limited, internal_error, service_unavailable, not_configured, forbidden, request_rejected, slot_taken, slot_unavailable, booking_limit_reached, idempotency_key_reused, not_signed_in, session_expired, sign_in_failed, account_locked, csrf_failed, demo_read_only, password_change_required, weak_password, email_taken, last_admin, transition_not_allowed, booking_changed, undo_unavailable] }
        message: { type: string }
        requestId: { type: string }
        retryAfterSeconds: { type: integer }
        details:
          type: array
          items: { type: object, required: [field, issue], properties: { field: { type: string }, issue: { type: string } } }
...
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { $ref: "#/components/schemas/ErrorInfo" } }
    Role: { type: string, enum: [admin, receptionist] }
    BookingStatus: { type: string, enum: [confirmed, arrived, completed, no_show, cancelled] }
    Viewer:
      type: object
      required: [kind, csrfToken, clinicToday, timezone]
      properties:
        kind: { type: string, enum: [staff, demo] }
        role: { $ref: "#/components/schemas/Role", description: "staff only" }
        displayName: { type: string }
        mustChangePassword: { type: boolean }
        csrfToken: { type: string }
        clinicToday: { type: string, format: date, description: "staff: today in clinic time; demo: the session's demo date" }
        timezone: { type: string, example: Asia/Karachi }
        sessionExpiresAt: { type: string, format: date-time, description: "earlier of idle and absolute expiry" }
        demoNow: { type: string, format: date-time, description: "demo only: the instant the demo treats as now" }
        typicalDay: { type: boolean, description: "demo only: true when outside clinic hours, so a typical clinic day is shown" }
    SignInRequest:
      type: object
      required: [email, password]
      properties: { email: { type: string, maxLength: 254 }, password: { type: string, maxLength: 128 } }
    SessionIssued:
      description: Returned to the website server only. The BFF moves `token` into the cookie and strips it from the browser response.
      type: object
      required: [token, viewer]
      properties: { token: { type: string }, viewer: { $ref: "#/components/schemas/Viewer" } }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties: { currentPassword: { type: string, maxLength: 128 }, newPassword: { type: string, minLength: 12, maxLength: 128 } }
...
    DoctorRef:
      type: object
      required: [id, name, departmentName]
      properties: { id: { type: string, format: uuid }, name: { type: string }, departmentName: { type: string }, isActive: { type: boolean } }
    BookingSummary:
      type: object
      required: [reference, startsAt, endsAt, localDate, localTime, status, version, patientNameMasked, phoneMasked, doctor, allowedNext]
      properties:
        reference: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        localDate: { type: string, format: date }
        localTime: { type: string }
        status: { $ref: "#/components/schemas/BookingStatus" }
        version: { type: integer }
        patientNameMasked: { type: string, example: "Ayesha K." }
        phoneMasked: { type: string, example: "0300****567" }
        doctor: { $ref: "#/components/schemas/DoctorRef" }
        allowedNext: { type: array, items: { $ref: "#/components/schemas/BookingStatus" }, description: "server-computed with the clock (FR-023/024)" }
        isSample: { type: boolean }
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/BookingSummary"
        - type: object
          required: [patientName, feePkr, bookedAt, history]
          properties:
            patientName: { type: string }
            emailMasked: { type: [string, "null"] }
            reason: { type: [string, "null"] }
            patientAge: { type: integer, description: "Pediatrics (demo sample data): the child's age" }
            bookedBy: { type: string, description: "Pediatrics (demo sample data): mother or father" }
            feePkr: { type: integer }
            bookedAt: { type: string, format: date-time }
            history:
              type: array
              items:
                type: object
                required: [at, toStatus, actor, isUndo]
                properties:
                  at: { type: string, format: date-time }
                  fromStatus: { oneOf: [ { $ref: "#/components/schemas/BookingStatus" }, { type: "null" } ] }
                  toStatus: { $ref: "#/components/schemas/BookingStatus" }
                  actor: { type: string, description: "staff display name or 'Online booking'" }
                  isUndo: { type: boolean }
    BookingSearchRequest:
      type: object
      properties:
        q: { type: string, maxLength: 80, description: reference or patient name; never logged }
        from: { type: string, format: date }
        to: { type: string, format: date }
        doctorId: { type: string, format: uuid }
        departmentId: { type: string, format: uuid }
        statuses: { type: array, items: { $ref: "#/components/schemas/BookingStatus" } }
        page: { type: integer, minimum: 1, default: 1 }
    BookingPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/BookingSummary" } }
        total: { type: integer }
        statusCounts: { type: object, additionalProperties: { type: integer }, description: "bookings per status for the same filters, ignoring the status filter (the chips)" }
        page: { type: integer }
        pageSize: { type: integer, const: 20 }
    StatusChangeRequest:
      type: object
      required: [to, expectedVersion]
      properties: { to: { $ref: "#/components/schemas/BookingStatus" }, expectedVersion: { type: integer } }
    StatusChangeResult:
      type: object
      required: [booking, changeId, undoExpiresAt]
      properties:
        booking: { $ref: "#/components/schemas/BookingDetail" }
        changeId: { type: string, format: uuid }
        undoExpiresAt: { type: string, format: date-time }
    UndoRequest:
      type: object
      required: [changeId]
      properties: { changeId: { type: string, format: uuid } }
    BookingChanged:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorInfo" }
        latest: { $ref: "#/components/schemas/BookingDetail" }
    PhoneReveal:
      type: object
      required: [phone, telHref, maskAfterSeconds]
      properties: { phone: { type: string, example: "0300 1234567" }, telHref: { type: string }, maskAfterSeconds: { type: integer, const: 60 } }
...